feat(sdTZ-M): allow one REE to invoke multiple TEEs
Generalize the sdTZ-M mechanism to support a 1:N mapping between one REE and
multiple TrustZone-M software-defined TEEs.

The previous implementation only supported one TEE connected to one REE. This
change enables the REE to select which TEE should handle each service request.

Validated with one FreeRTOS REE connected to two mTower TEEs, where one TEE
handles the Bitcoin Wallet TA and the other handles the Hello World TA.

+++ src/sdees/sdTZM/arch/armv8m/sdtzm.c
@@ -26,28 +26,14 @@ void sdtzm_copy_args_call_done(struct vcpu* vcpu_dst, struct vcpu* vcpu_src, siz
 }
 
 /* CROSSCON TODO Not good */
-extern long sdtzm_handler(struct vcpu* vcpu, uint64_t fid);
+extern long sdtzm_handler(struct vcpu* vcpu, uint32_t fid);
 
 static long sdtzm_hvc_handler(struct vcpu* vcpu, uint32_t smc_fid)
 {
     long ret = -HC_E_FAILURE;
 
-    //struct vcpu* calling_vcpu = cpu()->vcpu;
-
     ret = sdtzm_handler(vcpu, smc_fid);
 
-    // if (calling_vcpu->vm->type == 0) { /* normal world */ //not needed for v8-m case
-    //     if (is_psci_fid(smc_fid)) {
-    //         /* CROSSCON TODO: signal trusted OS a PSCI event is comming up */
-    //         /* potentially handle core going to sleep */
-    //         return HC_E_SUCCESS;
-    //     } else {
-    //         /* CROSSCON TODO: If HVC call is for trusted OS */
-    //         ret = sdtzm_handler(vcpu, smc_fid);
-    //     }
-    // } else {
-    //     ret = sdtzm_handler(vcpu, smc_fid);
-    // }
 
     return ret;
 }
